+++ json_serializable/test/json_serializable_test.dart
@@ -173,4 +173,7 @@ const _expectedSchemaTests = {
   'JsonSchemaNullableTest',
   'JsonSchemaNestedTest',
   'JsonSchemaNonCollectionTest',
+  'JsonSchemaGetterTest',
+  'JsonSchemaRecursiveListTest',
+  'JsonSchemaRecursiveListIssue',
 };